refactor(match-engine): universal SystemOutput protocol, float32 condition, shot pipeline, and test migration
BREAKING: No system mutates SimState directly — all return SystemOutput list
- SystemOutput protocol with FrameWrite DU for all frame mutations
- MatchStepper.applyOutputs as single mutation point
- ActionSystem orchestrates ActionResolver + RefereeApplicator
- EventRouter: BallSecured → ActivateAction (fixes idle players with ball)

test: complete test suite rewrite
- Delete 23 broken old-test files (old Possession/BallAgent API)
- Migrate 6 pure-logic files to Layer2_Systems/Layer3_EdgeCases
- New E2E tests in Layer1_E2E (FlowTransition, GoalCycle, GKDistribution, SetPieceFlow, MatchInvariant)
- New ApplyOutputContractTests verifying every SystemOutput case
- Migrate Enginetests.fs + Worldtests.fs to current API
- StatisticalContracts runs 20 parallel matches via lazy shared runner

+++ FootballEngine.Client/Engine/Match/Types/PlayerFrame.fs
@@ -92,14 +92,14 @@ module CognitiveOutputFrame =
           DefensiveRole = Array.create n (byte DefensiveRole.Marker) }
 
 type ConditionFrame =
-    { Condition: byte[]
+    { Condition: float32[]
       CachedTargetX: float32[]
       CachedTargetY: float32[]
       CachedExecution: float32[] }
 
 module ConditionFrame =
     let init (n: int) (roster: PlayerRoster) (basePositions: Spatial[]) : ConditionFrame =
-        { Condition = Array.init n (fun i -> byte roster.Players[i].Condition)
+        { Condition = Array.init n (fun i -> float32 roster.Players[i].Condition)
           CachedTargetX = Array.init n (fun i -> float32 basePositions[i].X)
           CachedTargetY = Array.init n (fun i -> float32 basePositions[i].Y)
           CachedExecution = Array.create n 1.0f }
